UI: Simplifing the frontend's URL routes.
We used to prefix every API call with `/api`, and then get webpack to
proxy it through to the real server. But this was always going to be a
problem when we went into production, because in production, there's
no prefix.

I've stripped the prefix out, and we'll just let webpack proxy
everything to the backend.